Fixed issues after moving files
